Adding the enroll admin script and starting the FabricNetwork class
The enrollAdmin.js is just a plain nodejs script for enrolling the admin into the network.
The FabricNetwork class is suppossed to be a singleton that contains all the logic for the application to connect to the blockchain. In it I just added the connectToNetwork method, which connects an identity to the network and returns an object containing the contract, network and gateway.